models/deepfake_detection/requirements.txt
Flask
torch
torchvision
Pillow
numpy
opencv-python
